Add previous_uuid column to git_poll_journal (#11)
When a change in the remote repository's git refs is detected the UUID
of the previous poll attempt will be recorded. This facilitates
searching the git poll history for changes. Now a simple `IS NOT NULL`
filter will identify the poll attempt that found new/updated refs and
a recursive query will provide the specific changes.

Fixes #6